+++ packages/client/base/src/types/payment-method-management/index.ts
@@ -1,3 +1,4 @@
 export * from "./events";
 export * from "./OrderIntents";
+export * from "./PaymentMethodAgenticEnrollment";
 export * from "./CrossmintPaymentMethodManagementProps";