+++ packages/client/ui/react-native/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @crossmint/client-sdk-react-native-ui
 
+## 1.1.9
+
+### Patch Changes
+
+- Updated dependencies [e03e7b1]
+- Updated dependencies [560fbf0]
+- Updated dependencies [0792f9e]
+  - @crossmint/wallets-sdk@1.0.10
+  - @crossmint/client-sdk-react-base@2.0.12
+
+## 1.1.8
+
+### Patch Changes
+
+- Updated dependencies [11f2089]
+- Updated dependencies [60f31bd]
+- Updated dependencies [b9225cf]
+  - @crossmint/client-sdk-base@2.3.0
+  - @crossmint/wallets-sdk@1.0.9
+  - @crossmint/client-sdk-auth@1.3.6
+  - @crossmint/client-sdk-react-base@2.0.11
+  - @crossmint/common-sdk-auth@1.1.4
+
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies [0535c3a]
+- Updated dependencies [612d768]
+  - @crossmint/client-sdk-auth@1.3.5
+  - @crossmint/wallets-sdk@1.0.8
+  - @crossmint/client-sdk-react-base@2.0.10
+
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [3a91c39]
+  - @crossmint/client-sdk-auth@1.3.4
+  - @crossmint/client-sdk-react-base@2.0.9
+
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies [448d797]
+  - @crossmint/client-sdk-react-base@2.0.8
+
+## 1.1.4
+
+### Patch Changes
+
+- @crossmint/wallets-sdk@1.0.7
+- @crossmint/client-sdk-react-base@2.0.7
+
 ## 1.1.3
 
 ### Patch Changes